get more results

+++ cmd/train_remake_model.py
@@ -25,6 +25,7 @@ sys.path.append(PATH_ROOT)
 
 TRAIN_DATASET_PATH = os.getenv('TRAIN_DATASET_PATH')
 model_path = os.path.join(PATH_ROOT, 'model', 'cnn_model.h5')
+retrained_model_path = os.path.join(PATH_ROOT, 'model', 'cnn_model_retrained.h5')
 
 from src.utils.shuffle import shuffle_in_order
 from src.utils.laod_retrain_dataset import load_retrain_dataset
@@ -40,48 +41,51 @@ retrained_images, retrained_labels = shuffle_in_order(retrained_images, retraine
 
 print('retrained_images.shape: ', retrained_images.shape, ', retrained_labels.shape: ', retrained_labels.shape)
 
-# Carrega o modelo salvo
-model = tf.keras.models.load_model(model_path)
+if os.path.exists(retrained_model_path):
+  pass
+else:
+  # Carrega o modelo salvo
+  model = tf.keras.models.load_model(model_path)
 
-# summary model structure
-model.summary() 
+  # summary model structure
+  model.summary() 
 
-# Compilando o modelo
-model.compile(optimizer='adam', loss='categorical_crossentropy', metrics=['accuracy'])
+  # Compilando o modelo
+  model.compile(optimizer='adam', loss='categorical_crossentropy', metrics=['accuracy'])
 
-# early stop
-callback = tf.keras.callbacks.EarlyStopping(monitor='loss', patience=0.075)
+  # early stop
+  callback = tf.keras.callbacks.EarlyStopping(monitor='loss', patience=0.075)
 
-# Supondo que seus rótulos são inteiros de 0 a 3
-retrained_labels_one_hot = tf.keras.utils.to_categorical(retrained_labels, num_classes=4)
+  # Supondo que seus rótulos são inteiros de 0 a 3
+  retrained_labels_one_hot = tf.keras.utils.to_categorical(retrained_labels, num_classes=4)
 
-# Treinando o modelo
-history = model.fit(retrained_images, retrained_labels_one_hot, batch_size=8, epochs=10, validation_split=0.1, callbacks=[callback], verbose=1)
+  # Treinando o modelo
+  history = model.fit(retrained_images, retrained_labels_one_hot, batch_size=8, epochs=10, validation_split=0.1, callbacks=[callback], verbose=1)
 
-# Plotando a acurácia de treino e validação ao longo das épocas
-plt.figure(figsize=(12, 5))
-plt.plot(history.history['accuracy'], label='Train Accuracy')
-plt.plot(history.history['val_accuracy'], label='Validation Accuracy')
-plt.legend()
-plt.title('Accuracy over epochs')
-plt.xlabel('Epoch')
-plt.ylabel('Accuracy')
+  # Plotando a acurácia de treino e validação ao longo das épocas
+  plt.figure(figsize=(12, 5))
+  plt.plot(history.history['accuracy'], label='Train Accuracy')
+  plt.plot(history.history['val_accuracy'], label='Validation Accuracy')
+  plt.legend()
+  plt.title('Accuracy over epochs')
+  plt.xlabel('Epoch')
+  plt.ylabel('Accuracy')
 
-plt.savefig(os.path.join(PATH_ROOT, 'stats', 'train', 'retrained_accuracy_graph.png'))
+  plt.savefig(os.path.join(PATH_ROOT, 'stats', 'train', 'retrained_accuracy_graph.png'))
 
-# Plotando a perda de treino e validação ao longo das épocas
-plt.figure(figsize=(12, 5))
-plt.plot(history.history['loss'], label='Train Loss')
-plt.plot(history.history['val_loss'], label='Validation Loss')
-plt.legend()
-plt.title('Loss over epochs')
-plt.xlabel('Epoch')
-plt.ylabel('Loss')
+  # Plotando a perda de treino e validação ao longo das épocas
+  plt.figure(figsize=(12, 5))
+  plt.plot(history.history['loss'], label='Train Loss')
+  plt.plot(history.history['val_loss'], label='Validation Loss')
+  plt.legend()
+  plt.title('Loss over epochs')
+  plt.xlabel('Epoch')
+  plt.ylabel('Loss')
 
-# save fig
-plt.savefig(os.path.join(PATH_ROOT, 'stats', 'train', 'retrained_loss_graph.png'))
+  # save fig
+  plt.savefig(os.path.join(PATH_ROOT, 'stats', 'train', 'retrained_loss_graph.png'))
 
-# Se você quiser salvar o modelo após o treinamento, você pode fazer isso:
-model.save(os.path.join(PATH_ROOT, 'model', 'cnn_model_retrained.h5'))
+  # Se você quiser salvar o modelo após o treinamento, você pode fazer isso:
+  model.save(os.path.join(PATH_ROOT, 'model', 'cnn_model_retrained.h5'))
 
 mlflow.run(uri='.', entry_point='retest_model_retrained', parameters={ 'tax_samples': tax_samples })
